requirements.txt
pymongo
flask
flask_login
python-dotenv
plotly